+++ src/TTX.Domain/ValueObjects/CreatorRarity.cs
@@ -7,9 +7,9 @@ public readonly struct CreatorRarity
     public required Creator Creator { get; init; }
     public required Rarity Rarity { get; init; }
 
-    public static CreatorRarity Create(long sum, Creator creator)
+    public static CreatorRarity Create(double sum, Creator creator)
     {
-        double calc = creator.Value / (double)sum * 100;
+        double calc = creator.Value / sum * 100.0;
         Rarity rarity = calc switch
         {
             >= 0 and < 1 => Rarity.Pennies,
